fix: ensure management socket is closed on failure to prevent a leak (#12)
I found this via dumb luck as I managed to knock a bluetooth adapter just out of the usb socket so it keeps disconnecting and reconnecting. Net results is a leak in python-btsocket which results in the bluetooth management socket not being closed if the stack doesn't respond so it leaves it open when it tries to reset it and leaks. Worse is the leak builds up over time if it happens again and if you have a busy systems its processing all the data while waiting for a response.

Make BluetoothMGMTProtocol a context manger and an asyncio.Protocol to ensure if anything goes wrong the underlying bluetooth management socket gets closed.